EHR 2.0 Disease prediction/EHR 2.0 Disease prediction/Server/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.2
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2